Change project folder structure to fit maven
so every homework would just be a maven module, and they can share some
code if necessary

+++ scripts/vagrant/jdk_setup.sh
@@ -1,22 +1,18 @@
-# Worskspace
+#!/usr/bin/env bash
 
-Code and dev environment setup scripts
+cd ~
+curl -LO 'http://download.oracle.com/otn-pub/java/jdk/7u79-b15/jdk-7u79-linux-x64.tar.gz' -H 'Cookie: oraclelicense=accept-securebackup-cookie'
 
-NOTE: the `Vagrantfile` doesn't config the sh file
-to install dependencies
+tarball="jdk-7u79-linux-x64.tar.gz"
 
-## Config java
+if [ ! -f "$tarball" ]; then
+    echo "jdk tar.gz file not found, exit"
+    exit
+fi
 
-- Download jdk7
-
-`curl -LO 'http://download.oracle.com/otn-pub/java/jdk/7u79-b15/jdk-7u79-linux-x64.tar.gz' -H 'Cookie: oraclelicense=accept-securebackup-cookie'`
-
-- extract and set java home
-
-````
 echo "extrcting jdk"
 sudo mkdir /usr/lib/jvm
-sudo tar zxvf ./jdk-7u79-linux-x64.tar.gz  -C /usr/lib/jvm
+sudo tar zxvf ${tarball}  -C /usr/lib/jvm
 cd /usr/lib/jvm
 sudo mv jdk1.7.0_79/ java-7-sun
 
@@ -36,10 +32,3 @@ sudo update-alternatives --install /usr/bin/javap javap /usr/lib/jvm/java-7-sun/
 
 echo "now java is working"
 java -version
-````
-
-You can try the `setup_jdk.sh` in scripts folder
-
-## Config hadoop
-
-`wget http://mirrors.cnnic.cn/apache/hadoop/common/stable/hadoop-2.7.1.tar.gz`
